Rename new methods to default
- Removes 0x.js dependency
- Sets default fee rate at 0.25%
- Renames order methods for v2
