+++ package.json
@@ -35,7 +35,6 @@
   "homepage": "https://github.com/ethfinex/efx-api-node",
   "license": "MIT",
   "dependencies": {
-    "0x.js": "^0.38.4",
     "@0x/order-utils": "^7.0.2",
     "aware": "^0.3.1",
     "bignumber.js": "^7.2.1",
