+++ src/api/contract/create_order.js
@@ -1,8 +1,15 @@
-const {ZeroEx} = require('0x.js')
+const {assetDataUtils, generatePseudoRandomSalt} = require('@0x/order-utils')
 const BigNumber = require('bignumber.js');
 
-module.exports = (efx, symbol, amount, price, validFor) => {
+module.exports = (efx, symbol, amount, price, validFor, fee_rate = 0.0025) => {
   const { web3, config } = efx
+
+  if(fee_rate < 0 || fee_rate > 0.05 || isNaN(fee_rate)){
+    // use 0.0025 for 0.25% fee
+    // use 0.01   for 1% fee
+    // use 0.05   for 5% fee
+    throw('fee_rate should be between 0 and 0.05')
+  }
 
   // symbols are always 3 letters
   const symbolOne = symbol.substr(0, symbol.length - 3)
@@ -17,21 +24,42 @@ module.exports = (efx, symbol, amount, price, validFor) => {
   let buyAmount, sellAmount
 
   if (amount > 0) {
-    buyAmount = (new BigNumber(10)).pow(buyCurrency.decimals).times(amount).times(1 + (buyCurrency.settleSpread || 0)).integerValue(BigNumber.ROUND_FLOOR).toString()
-    sellAmount = (new BigNumber(10)).pow(sellCurrency.decimals).times(amount).times(price).times(1 + (sellCurrency.settleSpread || 0)).integerValue(BigNumber.ROUND_FLOOR).toString()
+    buyAmount = (new BigNumber(10))
+      .pow(buyCurrency.decimals)
+      .times(amount)
+      .times(1 + (buyCurrency.settleSpread || 0))
+      .times(1 - fee_rate)
+      .integerValue(BigNumber.ROUND_FLOOR)
+
+    sellAmount = (new BigNumber(10))
+      .pow(sellCurrency.decimals)
+      .times(amount)
+      .times(price)
+      .times(1 + (sellCurrency.settleSpread || 0))
+      .integerValue(BigNumber.ROUND_FLOOR)
 
     // console.log( "Buying " + amount + ' ' + buySymbol + " for: " + price + ' ' + sellSymbol )
   }
 
   if (amount < 0) {
-    buyAmount = (new BigNumber(10)).pow(buyCurrency.decimals).times(amount).times(price).abs().times(1 + (buyCurrency.settleSpread || 0)).integerValue(BigNumber.ROUND_FLOOR).toString()
-    sellAmount = (new BigNumber(10)).pow(sellCurrency.decimals).times(amount).abs().times(1 + (sellCurrency.settleSpread || 0)).integerValue(BigNumber.ROUND_FLOOR).toString()
+    buyAmount = (new BigNumber(10))
+      .pow(buyCurrency.decimals)
+      .times(amount)
+      .times(price)
+      .abs()
+      .times(1 + (buyCurrency.settleSpread || 0))
+      .times(1 - fee_rate)
+      .integerValue(BigNumber.ROUND_FLOOR)
+
+    sellAmount = (new BigNumber(10))
+      .pow(sellCurrency.decimals)
+      .times(amount)
+      .abs()
+      .times(1 + (sellCurrency.settleSpread || 0))
+      .integerValue(BigNumber.ROUND_FLOOR)
 
     // console.log( "Selling " + Math.abs(amount) + ' ' + sellSymbol + " for: " + price + ' ' + buySymbol )
   }
-
-  // console.log( "   buy amount: " + buyAmount + " " + buySymbol )
-  // console.log( "  sell amount: " + sellAmount + " " + sellSymbol )
 
   let expiration
   expiration = Math.round((new Date()).getTime() / 1000)
@@ -39,21 +67,29 @@ module.exports = (efx, symbol, amount, price, validFor) => {
 
   // create order object
   const order = {
-    expirationUnixTimestampSec: web3.utils.toBN(expiration).toString(10),
-    feeRecipient: efx.config['0x'].ethfinexAddress.toLowerCase(),
+    makerAddress: efx.get('account').toLowerCase(),
+    takerAddress: '0x0000000000000000000000000000000000000000',
 
-    maker: efx.get('account').toLowerCase(),
+    feeRecipientAddress: efx.config['0x'].ethfinexAddress.toLowerCase(),
+    senderAddress: efx.config['0x'].ethfinexAddress.toLowerCase(),
+
+    makerAssetAmount: sellAmount,
+
+    takerAssetAmount: buyAmount,
+
     makerFee: web3.utils.toBN('0'),
-    makerTokenAddress: sellCurrency.wrapperAddress.toLowerCase(),
-    makerTokenAmount: sellAmount,
 
-    salt: ZeroEx.generatePseudoRandomSalt(),
-    taker: efx.config['0x'].ethfinexAddress.toLowerCase(),
     takerFee: web3.utils.toBN('0'),
-    takerTokenAddress: buyCurrency.wrapperAddress.toLowerCase(),
-    takerTokenAmount: buyAmount,
 
-    exchangeContractAddress: efx.config['0x'].exchangeAddress.toLowerCase()
+    expirationTimeSeconds: new BigNumber(expiration),
+
+    salt: generatePseudoRandomSalt(),
+
+    makerAssetData: assetDataUtils.encodeERC20AssetData(sellCurrency.wrapperAddress.toLowerCase()),
+
+    takerAssetData: assetDataUtils.encodeERC20AssetData(buyCurrency.wrapperAddress.toLowerCase()),
+
+    exchangeAddress: efx.config['0x'].exchangeAddress.toLowerCase()
   }
 
   return order
